Extracts index tables into their own components
